ripper/requirements.txt
psycopg2
pyyaml
numpy

face_recognition

minio

gunicorn
flask
requests